+++ tailwind.config.js
@@ -80,6 +80,7 @@ module.exports = {
         "custom-green": {
           lighter: "#D8EECA",
           darker: "#278400",
+          hover: "#237600",
         },
         "custom-red": {
           lighter: "#F3E9E8",